requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
numpy
psycopg2-binary
